requirements.txt
torch
torchvision
torchaudio

numpy
pillow
tqdm
SoccerNet
opencv-python
matplotlib
tensorboard
albumentations
lsq-ellipse